[#147814101][#115383335][#115383311] batch update
Dictionary Replacement
======================

Dictionary was using `std::string` to store any kind of data.
It's ok to use for data deserialization but it's totally inconvenient
and non optimal to use it with the lua tables.

Introducing DataProxy.

DataProxy wraps `Json::Value` or `sol::table` and allows to manipulate
it directly. It opened broad perspectives for the code.
New features:
1. It's now possible to pass lua functions and userdata directly to C++
code, but it should also support string parsing using
`TranslatorBetween`.
2. `Json::Value` can be decoded into `sol::table` keeping types.
3. Json serialization now writes stored types as is.

___

New Factory Paradigm
====================

For any kind of item there should be a global lua factory.
For example:
- `camera` -- `camera.createAndAttach('orbit')` will create a camera.
- `emitter` -- `emitter.create('damage')` create damage emitter.

Plan is to migrate characters creation to this paradigm as well.

___

Other Changes
=============

Camera factory allowed to reimplement two cameras in code.
Old `CameraFactory` was deleted.

Several bugfixes and improvements:
1. Fixed script system not removing update listener on entity cleanup.
2. `LuaScriptSystem` now supports set up and tear down script hooks hooks.
3. `ScriptComponent` can now save the current context.
4. `LuaScriptSystem` now allows passing lua functions as `setupFunction`
and `tearDownFunction`.

There are more fixes, but it was during 6 month development. I should
make this commits smaller...